requirements.txt
pandas
numpy
scikit-learn
beautifulsoup4
streamlit
requests
lxml
tqdm
nltk
matplotlib
seaborn
PyYAML
loguru
xgboost
psycopg2-binary
sqlalchemy
-e .